Counter table - svinja_broj_vakcina

+++ src/main/java/agri/api/nosql/svinja_broj_vakcina/SvinjaBrojVakcinaModel.java
@@ -3,12 +3,12 @@ package agri.api.nosql.svinja_broj_vakcina;
 public class SvinjaBrojVakcinaModel {
 
     private String  tetovir_broj_svinje;
-    private int broj_vakcina;
+    private long broj_vakcina;
 
     public SvinjaBrojVakcinaModel() {
     }
 
-    public SvinjaBrojVakcinaModel(String tetovir_broj_svinje, int broj_vakcina) {
+    public SvinjaBrojVakcinaModel(String tetovir_broj_svinje, long broj_vakcina) {
         this.tetovir_broj_svinje = tetovir_broj_svinje;
         this.broj_vakcina = broj_vakcina;
     }
@@ -21,11 +21,11 @@ public class SvinjaBrojVakcinaModel {
         this.tetovir_broj_svinje = tetovir_broj_svinje;
     }
 
-    public int getBroj_vakcina() {
+    public long getBroj_vakcina() {
         return broj_vakcina;
     }
 
-    public void setBroj_vakcina(int broj_vakcina) {
+    public void setBroj_vakcina(long broj_vakcina) {
         this.broj_vakcina = broj_vakcina;
     }
 }
